feat: multichain move ae libs to protocol dir
