+++ src/composables/accounts.ts
@@ -9,13 +9,11 @@ import type {
 } from '@/types';
 import { tg } from '@/store/plugins/languages';
 import { PROTOCOL_AETERNITY, PROTOCOL_BITCOIN } from '@/constants';
-import {
-  getAccountNameToDisplay,
-} from '@/popup/utils';
-import { AeScan } from '@/lib/AeScan';
+import { getAccountNameToDisplay } from '@/popup/utils';
 import { STUB_ACCOUNT } from '@/constants/stubs';
 import { AE_FAUCET_URL } from '@/protocols/aeternity/config';
 import { buildSimplexLink } from '@/protocols/aeternity/helpers';
+import { AeScan } from '@/protocols/aeternity/libs/AeScan';
 
 export function useAccounts({ store }: IDefaultComposableOptions) {
   // TODO in the future the state of the accounts should be stored in this composable
