+++ src/lib/FramesConnection.ts
@@ -2,7 +2,7 @@ import { times } from 'lodash-es';
 import { BrowserWindowMessageConnection } from '@aeternity/aepp-sdk';
 import { executeAndSetInterval } from '@/utils';
 import { handleUnknownError } from '@/popup/utils';
-import { AeSdkSupehero } from './AeSdkSupehero';
+import { AeSdkSuperhero } from '@/protocols/aeternity/libs/AeSdkSuperhero';
 
 const POLLING_INTERVAL = 3000;
 
@@ -22,7 +22,7 @@ export const FramesConnection = (() => {
     ];
   }
 
-  function init(aeSdk: AeSdkSupehero) {
+  function init(aeSdk: AeSdkSuperhero) {
     initialized = true;
     clearInterval(baseIntervalId);
